task2

+++ Tyuiu.GurevskayaVE.Sprint3.Task2.V25.Lib/DataService.cs
@@ -9,7 +9,8 @@ namespace Tyuiu.GurevskayaVE.Sprint3.Task2.V25.Lib
 {
     public class DataService : ISprint3Task2V25
     {
-        public double GetSumSeries(double value, int startValue, int stopValue)
+
+        public double GetSumSeries(int value, int startValue, int stopValue)
         {
             double SumSeries = 0;
             do
